+++ src/app/controller/home_resources.py
