+++ src/app/ext/auth.py
@@ -1,4 +1,5 @@
 from flask_jwt_extended import JWTManager
 
+
 def init_app(app):
     JWTManager(app)
