+++ src/app/ext/cors.py
@@ -1,4 +1,5 @@
 from flask_cors import CORS
 
+
 def init_app(app):
     CORS(app)
