+++ src/app/model/__init__.py
@@ -1,0 +1,6 @@
+from . import dao
+
+
+def init_app(app):
+    dao.init_app(app)
+
