+++ src/app/model/dao/__init__.py
@@ -1,0 +1,5 @@
+
+
+def init_app(app):
+    from . import course_dao, discipline_dao, post_dao, professor_dao, report_dao, student_dao
+
