+++ src/app/model/dao/report_dao.py
@@ -1,4 +1,4 @@
-from ..ext.database import db
+from ...ext.database import db
 import enum
