--- src/app/models/__init__.py
@@ -1,5 +1,0 @@
-
-
-def init_app(app):
-    from . import course_model, discipline_model, post_model, professor_model, report_model, student_model
-
