+++ src/app/view/__init__.py
@@ -5,3 +5,4 @@ ma = Marshmallow()
 
 def init_app(app):
     ma.init_app(app)
+
