+++ src/app/view/course_schema.py
@@ -1,11 +1,11 @@
 from . import ma
 from marshmallow import fields
-from app.models import course_model
+from ..model.dao import course_dao
 
 
 class CourseSchema(ma.SQLAlchemySchema):
     class Meta:
-        model = course_model.Course
+        model = course_dao.Course
 
     id_course = fields.Integer(required=True)
     name = fields.String(required=True)
